moved database out of services & rename repository package
