+++ database/database.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"log"
 
-	repository "github.com/piquel-fr/api/database/generated"
+	"github.com/piquel-fr/api/database/repository"
 	"github.com/piquel-fr/api/services/config"
 	"github.com/jackc/pgx/v5/pgxpool"
 )
