+++ main.go
@@ -5,10 +5,10 @@ import (
 	"log"
 	"net/http"
 
+	"github.com/piquel-fr/api/database"
 	"github.com/piquel-fr/api/handlers"
 	"github.com/piquel-fr/api/services/auth"
 	"github.com/piquel-fr/api/services/config"
-	"github.com/piquel-fr/api/services/database"
 	"github.com/piquel-fr/api/services/docs"
 	gh "github.com/piquel-fr/api/services/github"
 	"github.com/piquel-fr/api/services/middleware"
