+++ sqlc.yml
@@ -6,7 +6,7 @@ sql:
     gen:
       go:
         package: "repository"
-        out: "database/generated"
+        out: "database/repository"
         sql_package: "pgx/v5"
         emit_json_tags: true
         overrides:
